+++ modules/frontweb/src/main/java/com/example/frontweb/app/IndexController.java
@@ -1,4 +1,4 @@
-package com.example.frontweb.controller;
+package com.example.frontweb.app;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
