+++ modules/frontweb/src/main/java/com/example/frontweb/app/todo/TodoForm.java
@@ -1,4 +1,4 @@
-package com.example.frontweb.controller.todo;
+package com.example.frontweb.app.todo;
 
 import java.io.Serializable;
 
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class TodoForm implements Serializable {
     private static final long serialVersionUID = 1L;
-    
+
     private String todoId;
     private String todoTitle;
     private long version;
