+++ modules/frontweb/src/main/resources/application.properties
@@ -1,12 +1,14 @@
 spring.application.name=frontweb
 server.port=8080
-
 spring.main.banner-mode=off
+
+## ログレベル
+logging.level.com.example=DEBUG
 
 ## thymeleaf
 spring.thymeleaf.mode=HTML
 
-### registration
+## Spring Security (OAuth2クライアント)
 spring.security.oauth2.client.registration.keycloak.provider=keycloak
 spring.security.oauth2.client.registration.keycloak.client-id=demoapp
 spring.security.oauth2.client.registration.keycloak.client-secret=08c33835-c18c-4dd7-a7df-aee3479d17c4
@@ -18,5 +20,3 @@ spring.session.store-type=redis
 spring.redis.host=localhost
 spring.redis.port=6379
 
-## ログレベル
-logging.level.com.example=DEBUG
