--- modules/hello/src/main/java/com/example/hello/config/WebConfig.java
@@ -1,9 +1,0 @@
-package com.example.hello.config;
-
-import org.springframework.web.reactive.config.EnableWebFlux;
-import org.springframework.web.reactive.config.WebFluxConfigurer;
-
-@EnableWebFlux
-public class WebConfig implements WebFluxConfigurer {
-    
-}
