+++ modules/hello/src/main/resources/application.properties
@@ -1,5 +1,10 @@
-spring.main.banner-mode=off
 spring.application.name=hello
 server.port=8081
+spring.main.banner-mode=off
 
+## ログレベル
+logging.level.com.example=DEBUG
+
+## Spring Security (リソースサーバー)
 spring.security.oauth2.resourceserver.jwt.issuer-uri=http://localhost:18080/auth/realms/demo
+
