+++ modules/notifier/src/main/resources/application.properties
@@ -1,5 +1,8 @@
 spring.application.name=notifier
 spring.main.banner-mode=off
 
+## ログレベル
+logging.level.com.example=DEBUG
+
 ## ActiveMQ
 spring.activemq.broker-url=tcp://localhost:61616
