+++ modules/todo/src/main/java/com/example/todo/common/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.todo.config;
+package com.example.todo.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
@@ -10,13 +10,15 @@ public class SecurityConfig {
     @Bean
     public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http) {
         // 認可設定
+        // @formatter:off
         http.authorizeExchange()
                 .pathMatchers("/health").permitAll()
                 .pathMatchers("/message/**").hasAuthority("SCOPE_message:read")
                 .anyExchange().authenticated();
+        // @formatter:on
+
         // リソースサーバ
-        http.oauth2ResourceServer()
-                .jwt();
+        http.oauth2ResourceServer().jwt();
         return http.build();
     }
 }
