+++ modules/todo/src/main/java/com/example/todo/common/filter/WebLoggingFilter.java
@@ -1,4 +1,4 @@
-package com.example.frontweb.common;
+package com.example.todo.common.filter;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
@@ -13,22 +13,17 @@ import reactor.core.publisher.Mono;
 public class WebLoggingFilter implements WebFilter {
     @Override
     public Mono<Void> filter(ServerWebExchange exchange, WebFilterChain chain) {
-        if (exchange.getRequest().getPath().pathWithinApplication().value().startsWith("/css/")) {
-            return chain.filter(exchange);
-        }
-
         return chain.filter(exchange).transformDeferred(call -> Mono.fromRunnable(() -> {
             // Before
             var req = exchange.getRequest();
             log.info("[WEB]REQUEST: {} {}", req.getMethod(), req.getURI());
-            // log.info("[WEB]REQUEST cookies={}", req.getCookies());
-            // log.info("[WEB]REQUEST headers={}", req.getHeaders());
+            log.info("[WEB]REQUEST headers={}", req.getHeaders());
         }).then(call).doOnSuccess(done -> {
-            // after (success)
+            // After (success)
             var rsp = exchange.getResponse();
             log.info("[WEB]SUCCESS: statusCode={}", rsp.getStatusCode());
         }).doOnError(throwable -> {
-            // after (with error)
+            // After (error)
             log.info("[WEB]ERROR: {}", throwable, throwable.getMessage());
         }));
     }
