--- modules/todo/src/main/java/com/example/todo/config/WebConfig.java
@@ -1,9 +1,0 @@
-package com.example.todo.config;
-
-import org.springframework.web.reactive.config.EnableWebFlux;
-import org.springframework.web.reactive.config.WebFluxConfigurer;
-
-@EnableWebFlux
-public class WebConfig implements WebFluxConfigurer {
-    
-}
