+++ modules/todo/src/main/java/com/example/todo/domain/service/TodoServiceImpl.java
@@ -8,13 +8,11 @@ import com.example.todo.domain.repository.TodoRepository;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @Service
 @RequiredArgsConstructor
-@Slf4j
 public class TodoServiceImpl implements TodoService {
     private final TodoRepository todoRepository;
 
@@ -26,8 +24,6 @@ public class TodoServiceImpl implements TodoService {
             entity.setVersion(0);
         }
 
-        log.info("★save {}", entity);
-        
         return todoRepository.save(entity);
     }
