+++ modules/todo/src/main/resources/application.properties
@@ -1,18 +1,18 @@
 spring.application.name=todo
 server.port=8082
-
 spring.main.banner-mode=off
 
-## Security
+## ログレベル
+logging.level.com.example=DEBUG
+logging.level.org.springframework.data.r2dbc=DEBUG
+
+## Spring Security (OAuth2リソースサーバー)
 spring.security.oauth2.resourceserver.jwt.issuer-uri=http://localhost:18080/auth/realms/demo
 
-## R2DBC
+## Spring Data R2DBC
 spring.r2dbc.url=r2dbc:pool:postgresql://localhost:5432/postgres
 spring.r2dbc.username=postgres
 spring.r2dbc.password=postgres
-
-## R2DBC(ログ)
-logging.level.org.springframework.data.r2dbc=DEBUG
 
 ## ActiveMQ
 spring.activemq.broker-url=tcp://localhost:61616
